+++ bt/actions/bt_fail.cpp
@@ -2,6 +2,6 @@
 
 #include "bt_fail.h"
 
-int BTFail::_tick(float p_delta) {
+int BTFail::_tick(double p_delta) {
 	return FAILURE;
 }
