+++ bt/actions/bt_wait.h
@@ -10,23 +10,23 @@ class BTWait : public BTAction {
 	GDCLASS(BTWait, BTAction);
 
 private:
-	float duration = 1.0;
+	double duration = 1.0;
 
-	float time_passed = 0.0;
+	double time_passed = 0.0;
 
 protected:
 	static void _bind_methods();
 
 	virtual String _generate_name() const override;
 	virtual void _enter() override;
-	virtual int _tick(float p_delta) override;
+	virtual int _tick(double p_delta) override;
 
 public:
-	void set_duration(float p_value) {
+	void set_duration(double p_value) {
 		duration = p_value;
 		emit_changed();
 	}
-	float get_duration() const { return duration; }
+	double get_duration() const { return duration; }
 };
 
 #endif // BT_WAIT_H
