+++ bt/composites/bt_random_selector.cpp
@@ -13,7 +13,7 @@ void BTRandomSelector::_enter() {
 	indicies.shuffle();
 }
 
-int BTRandomSelector::_tick(float p_delta) {
+int BTRandomSelector::_tick(double p_delta) {
 	int status = FAILURE;
 	for (int i = last_running_idx; i < get_child_count(); i++) {
 		status = get_child(indicies[i])->execute(p_delta);
