+++ bt/decorators/bt_time_limit.h
@@ -10,22 +10,22 @@ class BTTimeLimit : public BTDecorator {
 	GDCLASS(BTTimeLimit, BTDecorator);
 
 private:
-	float time_limit = 5.0;
-	float time_passed = 0.0;
+	double time_limit = 5.0;
+	double time_passed = 0.0;
 
 protected:
 	static void _bind_methods();
 
 	virtual String _generate_name() const override;
 	virtual void _enter() override;
-	virtual int _tick(float p_delta) override;
+	virtual int _tick(double p_delta) override;
 
 public:
-	void set_time_limit(float p_value) {
+	void set_time_limit(double p_value) {
 		time_limit = p_value;
 		emit_changed();
 	}
-	float get_time_limit() const { return time_limit; }
+	double get_time_limit() const { return time_limit; }
 };
 
 #endif // BT_TIME_LIMIT_H
