+++ limbo_hsm.h
@@ -38,7 +38,7 @@ protected:
 	virtual void _initialize(Node *p_agent, const Ref<Blackboard> &p_blackboard) override;
 	virtual void _enter() override;
 	virtual void _exit() override;
-	virtual void _update(float p_delta) override;
+	virtual void _update(double p_delta) override;
 
 	void _change_state(LimboState *p_state);
 
@@ -56,7 +56,7 @@ public:
 	virtual void initialize(Node *p_agent, const Ref<Blackboard> &p_parent_scope = nullptr);
 	virtual bool dispatch(const String &p_event, const Variant &p_cargo) override;
 
-	void update(float p_delta) { _update(p_delta); }
+	void update(double p_delta) { _update(p_delta); }
 	void add_transition(Node *p_from_state, Node *p_to_state, const String &p_event);
 	// void add_transition_from_any_state(Node *p_to_state, const String &p_event);
 	LimboState *anystate() const { return nullptr; };
